Simplify CLI schema and move functions to definition.py
- Change CLI schema from object with 'function' property to simple string array
- Update ProvidersManager to parse CLI items as strings directly
- Move get_xxx_cli_commands functions from __init__.py to definition.py
- Update all provider.yaml files to use simplified format: `cli: [path.to.function]`
