+++ airflow-core/src/airflow/provider_info.schema.json
@@ -359,16 +359,7 @@
             "type": "array",
             "description": "CLI command functions exposed by the provider",
             "items": {
-                "type": "object",
-                "properties": {
-                    "function": {
-                        "type": "string",
-                        "description": "Python path to the function that returns CLI commands"
-                    }
-                },
-                "required": [
-                    "function"
-                ]
+                "type": "string"
             }
         },
         "config": {
