+++ providers/amazon/provider.yaml
@@ -1300,7 +1300,7 @@ auth-managers:
   - airflow.providers.amazon.aws.auth_manager.aws_auth_manager.AwsAuthManager
 
 cli:
-  - function: airflow.providers.amazon.aws.auth_manager.cli.get_aws_cli_commands
+  - airflow.providers.amazon.aws.auth_manager.cli.definition.get_aws_cli_commands
 
 queues:
   - airflow.providers.amazon.aws.queues.sqs.SqsMessageQueueProvider
