+++ providers/fab/provider.yaml
@@ -254,5 +254,5 @@ auth-managers:
   - airflow.providers.fab.auth_manager.fab_auth_manager.FabAuthManager
 
 cli:
-  - function: airflow.providers.fab.auth_manager.cli_commands.get_fab_cli_commands
+  - airflow.providers.fab.auth_manager.cli_commands.definition.get_fab_cli_commands
