+++ providers/fab/src/airflow/providers/fab/auth_manager/cli_commands/definition.py
@@ -336,3 +336,32 @@ DB_COMMANDS = (
         args=(ARG_YES, ARG_DB_SKIP_INIT, ARG_VERBOSE),
     ),
 )
+
+
+def get_fab_cli_commands():
+    """Return CLI commands for FAB auth manager."""
+    import packaging.version
+
+    from airflow import __version__ as airflow_version
+    from airflow.cli.cli_config import GroupCommand
+
+    commands = [
+        GroupCommand(
+            name="users",
+            help="Manage users",
+            subcommands=USERS_COMMANDS,
+        ),
+        GroupCommand(
+            name="roles",
+            help="Manage roles",
+            subcommands=ROLES_COMMANDS,
+        ),
+        SYNC_PERM_COMMAND,  # not in a command group
+        PERMISSIONS_CLEANUP_COMMAND,  # single command for permissions cleanup
+    ]
+    # If Airflow version is 3.0.0 or higher, add the fab-db command group
+    if packaging.version.parse(
+        packaging.version.parse(airflow_version).base_version
+    ) >= packaging.version.parse("3.0.0"):
+        commands.append(GroupCommand(name="fab-db", help="Manage FAB", subcommands=DB_COMMANDS))
+    return commands
