+++ providers/keycloak/provider.yaml
@@ -36,7 +36,7 @@ auth-managers:
   - airflow.providers.keycloak.auth_manager.keycloak_auth_manager.KeycloakAuthManager
 
 cli:
-  - function: airflow.providers.keycloak.auth_manager.cli.get_keycloak_cli_commands
+  - airflow.providers.keycloak.auth_manager.cli.definition.get_keycloak_cli_commands
 
 config:
   keycloak_auth_manager:
